Lower non contiguous assumed-shape and CONTIGUOUS keyword
- Add a field in AbstractArrayBox to keep track of the fir.box
  that describes the array. Add a isContiguous helper to tell
  if the fir.box is known to be contiguous or not (looking at
  its type and attributes).

- Update symbol lowering to propagate dummy fir.box to created array SymbolBox.

- Update parts of lowering that generate fir.coordinate_of, fir.array_coor,
  and fir.array_load with such non contiguous entities to pass the fir.box
  to the fir operations instead of the bare address.

- Update the CallInterface to add a "fir.contribuous" attribute to the funcOp
  arguments of entities that have the CONTIGUOUS argument in Fortran.

- Update pointer/allocatable read to propagate the fir.box, if it exists, to
  the created AbstractArrayBox.

Note: I think that it would be nicer to make the base mlir::Value be the fir.box rather
than to add a new field for the fir.box. But that requires more refactoring and thinking
(like should we do it also for non-contiguous/scalar entities that happens to have a fir.box).
Few things making fir::getBase(exv) are currently expecting a fir.box.
A new exv.getMemRef() is added, it returns the fir.box of discontiguous entity,
or the base address otherwise. It is intended to centralized things to help
migrating towards having fir.box base mlir::Value in boxvalue if we want to.